+++ wakamiti-engine/wakamiti-core/src/main/java/es/iti/wakamiti/core/datatypes/assertion/UnaryNumberAssertProvider.java
@@ -6,7 +6,7 @@
 package es.iti.wakamiti.core.datatypes.assertion;
 
 
-import es.iti.wakamiti.core.backend.ExpressionMatcher;
+import es.iti.wakamiti.api.ExpressionMatcher;
 import org.hamcrest.Matcher;
 import org.hamcrest.Matchers;